+++ lib/jay_api/elasticsearch/errors.rb
@@ -6,6 +6,7 @@ require_relative 'errors/query_execution_error'
 require_relative 'errors/query_execution_failure'
 require_relative 'errors/query_execution_timeout'
 require_relative 'errors/search_after_error'
+require_relative 'errors/writable_index_error'
 
 module JayAPI
   module Elasticsearch